Move GUI tests to be under main library `gui` directory with own marker
