+++ pytest.ini
@@ -7,4 +7,4 @@ log_file_date_format = %Y-%m-%dT%H:%M:%S
 log_file_format = %(asctime)s %(levelname)-7s <%(filename)s:%(lineno)d> %(message)s
 log_file_level = DEBUG
 markers =
-    requires_human(action): Mark a test that requires some kind of non-automated action (e.g. closing a GUI window)
+    gui: Mark tests targeting the FOQUS graphical interface through the pytest-qt framework
